Backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
pymongo
python-dotenv
python-multipart
